bot/requirements.txt
redis==5.0.4
aiogram==3.5.0
python-dotenv==1.0.1
dogpile.cache==1.3.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
langchain
langchain_community
sentence-transformers
faiss-gpu
docx2txt

